+++ Create-IncrementalBackupJob.ps1
@@ -3,15 +3,25 @@
 
 #create incremental scheduled job
 
-import-module PSScheduledJob
+Import-Module PSScheduledJob
 
 $filepath = "C:\scripts\PSBackup\DailyIncrementalBackup.ps1"
 
 if (Test-Path $filepath) {
-    $trigger = New-JobTrigger -At 10:00PM -DaysOfWeek Saturday,Sunday,Monday,Tuesday,Wednesday,Thursday -Weekly
+    $trigger = New-JobTrigger -At 10:00PM -DaysOfWeek Saturday, Sunday, Monday, Tuesday, Wednesday, Thursday -Weekly
     $jobOpt = New-ScheduledJobOption -RunElevated -RequireNetwork -WakeToRun
 
-    Register-ScheduledJob -FilePath $filepath -Name "DailyIncremental" -Trigger $trigger -ScheduledJobOption $jobOpt -MaxResultCount 7 -Credential jeff
+    #parameters for Register-ScheduledJob
+    $job = @{
+        FilePath           = $filepath
+        Name               = "DailyIncremental"
+        Trigger            = $trigger
+        ScheduledJobOption = $jobOpt
+        MaxResultCount     = 7
+        Credential         = $env:username
+    }
+
+    Register-ScheduledJob @job
 }
 else {
     Write-Warning "Can't find $filepath"